fix: workaround for disable bug
implemented a workaround for a bug that causes saving the settings to disable the user storage provider and all users linked to it: https://github.com/keycloak/keycloak/issues/21045

also made a small workaround in onUpdated to ensure current periodic syncs are refreshed, so they get new settings, when saving. Be default it only happens if changes are made to the sync intervals.

implemented logging admin events when on start, finish and error.
this helps ensuring that the sync is running as expected.

made some performance improvements by running some stuff in parallel
and avoiding too small or too large transactions